+++ language_chinese.h
@@ -1,0 +1,298 @@
+#ifndef MENU_TEXT_MENU
+  #define MENU_TEXT_MENU                                "* 菜单                     # 按键定义"               // "* Menu # Key Defs"
+#endif
+#ifndef MENU_TEXT_MENU_HASH_IS_FUNCTIONS
+  #define MENU_TEXT_MENU_HASH_IS_FUNCTIONS              "* 菜单                         # Fn"          // "* Menu # Fn"
+#endif
+#ifndef MENU_TEXT_FINISH
+  #define MENU_TEXT_FINISH                              "                            # 完成"              // "# Finish"
+#endif
+#ifndef MENU_TEXT_CANCEL
+  #define MENU_TEXT_CANCEL                              "* 取消       "                                   // "* Cancel"
+#endif
+#ifndef MENU_TEXT_SHOW_DIRECT
+  #define MENU_TEXT_SHOW_DIRECT                         "                       # 直显按键功能"               // "# Show Direct"
+#endif
+#ifndef MENU_TEXT_ROSTER
+  #define MENU_TEXT_ROSTER                              "* 取消   0-9             #翻页"                    // "* Cancel 0-9 #Pg"
+#endif
+#ifndef MENU_TEXT_TURNOUT_LIST
+  #define MENU_TEXT_TURNOUT_LIST                        "* 取消   0-9             #翻页"                    // "* Cancel 0-9 #Pg"
+#endif
+#ifndef MENU_TEXT_ROUTE_LIST
+  #define MENU_TEXT_ROUTE_LIST                          "* 取消   0-9             #翻页"                    // "* Cancel 0-9 #Pg"
+#endif
+#ifndef MENU_TEXT_FUNCTION_LIST
+  #define MENU_TEXT_FUNCTION_LIST                       "* 取消   0-9             #翻页"                    // "* Cancel 0-9 #Pg"
+#endif
+#ifndef MENU_TEXT_SELECT_WIT_SERVICE
+  #define MENU_TEXT_SELECT_WIT_SERVICE                  "0-4       # 进入， 按下旋钮 关闭"             // "0-4 # Entry E.btn OFF"
+#endif
+#ifndef MENU_TEXT_SELECT_WIT_ENTRY
+  #define MENU_TEXT_SELECT_WIT_ENTRY                    "0-9  * 返回  #确认   按下旋钮关闭"                // "0-9 * Back # Go E.btn OFF"
+#endif
+#ifndef MENU_TEXT_SELECT_SSIDS
+  #define MENU_TEXT_SELECT_SSIDS                        "0-9         # 搜索        按下旋钮关闭"             // "0-9 # Search E.btn OFF"
+#endif
+#ifndef MENU_TEXT_SELECT_SSIDS_FROM_FOUND
+  #define MENU_TEXT_SELECT_SSIDS_FROM_FOUND             "0-4  9 列表  # 翻页  按下旋钮 关闭"                  // "0-4 9 List # Pg E.btn OFF"
+#endif
+#ifndef MENU_TEXT_ENTER_SSID_PASSWORD
+  #define MENU_TEXT_ENTER_SSID_PASSWORD                 "旋钮选字符    # 确认   * 返回"                  // "E Chrs E.btn Slct # Go * Bck"
+#endif
+
+
+#ifndef DIRECT_COMMAND_LIST
+  #define DIRECT_COMMAND_LIST                           "直接指令"                               // "Direct Commands"
+#endif
+#ifndef DIRECTION_FORWARD_TEXT
+  #define DIRECTION_FORWARD_TEXT                        "前进"                                          // "Fwd"
+#endif
+#ifndef DIRECTION_REVERSE_TEXT
+  #define DIRECTION_REVERSE_TEXT                        "后退"                                         // "Rev"
+#endif
+
+
+#ifndef MSG_START
+  #define MSG_START                                     "启动"                                         // "Start"
+#endif
+#ifndef MSG_BROWSING_FOR_SERVICE
+  #define MSG_BROWSING_FOR_SERVICE                      "搜索WiT服务"                       // "Browsing for WiT services"
+#endif
+#ifndef MSG_BROWSING_FOR_SSIDS
+  #define MSG_BROWSING_FOR_SSIDS                        "搜索无线网络"                              // "Browsing for SSIDs"
+#endif
+#ifndef MSG_NO_SSIDS_FOUND
+  #define MSG_NO_SSIDS_FOUND                            "未发现无线网络"                          // "No SSIDs found"
+#endif
+#ifndef MSG_SSIDS_LISTED
+  #define MSG_SSIDS_LISTED                              " 已知无线网络"                               // "SSIDs listed"
+#endif
+#ifndef MSG_SSIDS_FOUND
+  #define MSG_SSIDS_FOUND                               " 发现无线网络"                              // "SSIDs found"
+#endif
+#ifndef MSG_BOUNJOUR_SETUP_FAILED
+  #define MSG_BOUNJOUR_SETUP_FAILED                     "监听服务未启动"                    // "Unable to setup Listener"
+#endif
+#ifndef MSG_NO_SERVICES_FOUND
+  #define MSG_NO_SERVICES_FOUND                         "未发现服务"                        // "No services found"
+#endif
+#ifndef MSG_NO_SERVICES_FOUND_ENTRY_REQUIRED
+  #define MSG_NO_SERVICES_FOUND_ENTRY_REQUIRED          "输入WiT服务器IP:端口"                    // "Enter witServer IP:Port"
+#endif
+#ifndef MSG_SERVICES_FOUND
+  #define MSG_SERVICES_FOUND                            "发现服务器"                            // "Service(s) found"
+#endif
+#ifndef MSG_TRYING_TO_CONNECT
+  #define MSG_TRYING_TO_CONNECT                         "正在尝试连接"                         // "Trying to Connect"
+#endif
+#ifndef MSG_CONNECTED
+  #define MSG_CONNECTED                                 "             已连接"                        // "Connected"
+#endif
+#ifndef MSG_CONNECTING
+  #define MSG_CONNECTING                                "             连接中..."                     // "Connecting..."
+#endif
+
+
+#ifndef MSG_ADDRESS_LABEL
+  #define MSG_ADDRESS_LABEL                             "IP地址：                "                                   // "IP address:"
+#endif
+#ifndef MSG_CONNECTION_FAILED
+  #define MSG_CONNECTION_FAILED                         "连接失败"                         // "Connection failed"
+#endif
+#ifndef MSG_DISCONNECTED
+  #define MSG_DISCONNECTED                              "已断开"                                      // "Disconnected"
+#endif
+#ifndef MSG_AUTO_SLEEP
+  #define MSG_AUTO_SLEEP                                "选择超时"                 // "Waited too long for Select"
+#endif
+#ifndef MSG_BATTERY_SLEEP
+  #define MSG_BATTERY_SLEEP                             "电池电量低"                         // "Battery critically low"
+#endif
+#ifndef MSG_START_SLEEP
+  #define MSG_START_SLEEP                               "正在关机。      @ 重启"               // "Shutting Down. E.btn ON"
+#endif
+#ifndef MSG_THROTTLE_NUMBER
+  #define MSG_THROTTLE_NUMBER                           "             控制器 #"                         // "Throttle #"
+#endif
+#ifndef MSG_NO_LOCO_SELECTED
+  #define MSG_NO_LOCO_SELECTED                          "      未选择机车"                    // "No Loco selected"
+#endif
+#ifndef MSG_ENTER_PASSWORD
+  #define MSG_ENTER_PASSWORD                            "输入密码"                             // "Enter Password"
+#endif
+#ifndef MSG_GUESSED_EX_CS_WIT_SERVER
+  #define MSG_GUESSED_EX_CS_WIT_SERVER                  "发现EX-CS WiT服务器"                     // "Guessed' EX-CS WiT server"
+#endif
+#ifndef MSG_BYPASS_WIT_SERVER_SEARCH
+  #define MSG_BYPASS_WIT_SERVER_SEARCH                  "跳过WiT服务器搜索"                       // "Bypass WiT server search"
+#endif
+#ifndef MSG_NO_FUNCTIONS
+  #define MSG_NO_FUNCTIONS                              "功能列表 - 列表为空"                   // "Function List - No Functions"
+#endif
+#ifndef MSG_HEARTBEAT_CHECK_ENABLED
+  #define MSG_HEARTBEAT_CHECK_ENABLED                   "心跳检测已启用"                   // "Heartbeat Check Enabled"
+#endif
+#ifndef MSG_HEARTBEAT_CHECK_DISABLED
+  #define MSG_HEARTBEAT_CHECK_DISABLED                  "心跳检测已禁用"                 // "Heartbeat Check Disabled"
+#endif
+#ifndef MSG_RECEIVING_SERVER_DETAILS
+   #define MSG_RECEIVING_SERVER_DETAILS                 "正在接收服务器数据"                    // "Getting data from server"
+#endif
+
+#ifndef MENU_ITEM_TEXT_TITLE_FUNCTION
+  #define MENU_ITEM_TEXT_TITLE_FUNCTION                 "功能"                                      // "Function"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_ADD_LOCO
+  #define MENU_ITEM_TEXT_TITLE_ADD_LOCO                 "添加机车"                                   // "Add Loco"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_DROP_LOCO
+  #define MENU_ITEM_TEXT_TITLE_DROP_LOCO                "移除机车"                                     // "Drop Loco"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_TOGGLE_DIRECTION
+  #define MENU_ITEM_TEXT_TITLE_TOGGLE_DIRECTION         "切换方向"                                  // "Toggle Dir"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_SPEED_STEP_MULTIPLIER
+  #define MENU_ITEM_TEXT_TITLE_SPEED_STEP_MULTIPLIER    "速度倍率"                                  // "X Speed Step"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_THROW_POINT
+  #define MENU_ITEM_TEXT_TITLE_THROW_POINT              "道岔转辙"                                   // "Throw Point"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_CLOSE_POINT
+  #define MENU_ITEM_TEXT_TITLE_CLOSE_POINT              "道岔归位"                                   // "Close Point"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_ROUTE
+  #define MENU_ITEM_TEXT_TITLE_ROUTE                    "进路"                                         // "Route"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_TRACK_POWER
+  #define MENU_ITEM_TEXT_TITLE_TRACK_POWER              "轨道供电"                                   // "Trk Power"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_EXTRAS
+  #define MENU_ITEM_TEXT_TITLE_EXTRAS                   "附加功能"                                        // "Extras"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_HEARTBEAT
+  #define MENU_ITEM_TEXT_TITLE_HEARTBEAT                "心跳检测"                                    // "Heartbeat"
+#endif
+#ifndef MENU_ITEM_TEXT_TITLE_EDIT_CONSIST
+  #define MENU_ITEM_TEXT_TITLE_EDIT_CONSIST             "编辑重联"                            // "Edit Consist Facing"
+#endif
+
+
+#ifndef MENU_ITEM_TEXT_MENU_FUNCTION
+  #define MENU_ITEM_TEXT_MENU_FUNCTION                  "编号+# 选择  * 取消  # 列表"                  // "no+# Select * Cancel # List"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_ADD_LOCO
+  #define MENU_ITEM_TEXT_MENU_ADD_LOCO                  "地址+# 添加  * 取消  # 机车库"                 // "addr+# Add * Cancel # Roster"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_DROP_LOCO
+  #define MENU_ITEM_TEXT_MENU_DROP_LOCO                 "地址+# 单删   * 取消   # 全删"                  // "addr+# One * Cancel # All"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_DROP_LOCO_BYINDEX
+  #define MENU_ITEM_TEXT_MENU_DROP_LOCO_BY_INDEX        "编号+# 单删   * 取消   # 全删"                   // "no+# One * Cancel # All"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_TOGGLE_DIRECTION
+  #define MENU_ITEM_TEXT_MENU_TOGGLE_DIRECTION          "# 切换"                                  // "# Toggle"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_SPEED_STEP_MULTIPLIER
+  #define MENU_ITEM_TEXT_MENU_SPEED_STEP_MULTIPLIER     "* 取消      # 切换"                 // "* Cancel # Toggle"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_THROW_POINT
+  #define MENU_ITEM_TEXT_MENU_THROW_POINT               "编号+# 转辙  * 取消  # 列表"                 // "no+# Throw * Cancel # List"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_CLOSE_POINT
+  #define MENU_ITEM_TEXT_MENU_CLOSE_POINT               "编号+# 归位  * 取消  # 列表"                  // "no+# Close * Cancel # List"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_ROUTE
+  #define MENU_ITEM_TEXT_MENU_ROUTE                     "编号+# 选择 * 取消 # 列表"                 // "no+# Select * Cancel # List"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_TRACK_POWER
+  #define MENU_ITEM_TEXT_MENU_TRACK_POWER               "* 取消      # 切换"                 // "* Cancel # Toggle"
+  #endif
+#ifndef MENU_ITEM_TEXT_MENU_EXTRAS
+  #define MENU_ITEM_TEXT_MENU_EXTRAS                    "选择编号        * 取消"                  // "no Select * Cancel"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_HEARTBEAT
+  #define MENU_ITEM_TEXT_MENU_HEARTBEAT                 "* 关闭"                                   // "* Close"
+#endif
+#ifndef MENU_ITEM_TEXT_MENU_EDIT_CONSIST
+  #define MENU_ITEM_TEXT_MENU_EDIT_CONSIST              "编号 反转方向  * 关闭"               // "no Chng Facing * Close"
+#endif
+
+
+#ifndef EXTRA_MENU_TEXT_CHAR_FUNCTION_KEY_TOGGLE
+  #define EXTRA_MENU_TEXT_CHAR_FUNCTION_KEY_TOGGLE      "功能/按键 切换"                                // "Fnc/Key Tgl"
+#endif
+#ifndef EXTRA_MENU_TEXT_CHAR_EDIT_CONSIST
+  #define EXTRA_MENU_TEXT_CHAR_EDIT_CONSIST             "编辑重联"                                  // "Edt Consist"
+#endif
+
+#ifndef EXTRA_MENU_TEXT_CHAR_HEARTBEAT_TOGGLE
+  #define EXTRA_MENU_TEXT_CHAR_HEARTBEAT_TOGGLE         "心跳 开关"                                   // "Heartbt Tgl"
+#endif
+#ifndef EXTRA_MENU_TEXT_CHAR_DROP_BEFORE_ACQUIRE_TOGGLE
+  #define EXTRA_MENU_TEXT_CHAR_DROP_BEFORE_ACQUIRE_TOGGLE "单机车 开关"                                   // "1 loco Tgl"
+#endif
+#ifndef EXTRA_MENU_TEXT_SAVE_CURRENT_LOCOS
+  #define EXTRA_MENU_TEXT_SAVE_CURRENT_LOCOS            "保存当前机车"                                   // "Save Locos"
+#endif
+#ifndef EXTRA_MENU_TEXT_CHAR_INCREASE_MAX_THROTTLES
+  #define EXTRA_MENU_TEXT_CHAR_INCREASE_MAX_THROTTLES   "控制器数 +"                                     // "#Throttles +"
+#endif
+#ifndef EXTRA_MENU_TEXT_CHAR_DECREASE_MAX_THROTTLES
+  #define EXTRA_MENU_TEXT_CHAR_DECREASE_MAX_THROTTLES   "控制器数 -"                                     // "#Throttles -"
+#endif
+#ifndef EXTRA_MENU_TEXT_CHAR_DISCONNECT
+  #define EXTRA_MENU_TEXT_CHAR_DISCONNECT               "断开WiT连接"                                    // "Disconnect"
+#endif
+#ifndef EXTRA_MENU_TEXT_CHAR_OFF_SLEEP
+  #define EXTRA_MENU_TEXT_CHAR_OFF_SLEEP                "关机 / 休眠"                                  // "OFF / Sleep"
+#endif
+
+
+#ifndef F0_LABEL
+  #define F0_LABEL                                      "灯光"                                         // "Light"
+#endif
+#ifndef F1_LABEL
+  #define F1_LABEL                                      "铃铛"                                        // "Bell"
+#endif
+#ifndef F2_LABEL
+  #define F2_LABEL                                      "喇叭"                                          // "Horn"
+#endif
+
+#ifndef CHOSEN_KEYPAD_DISPLAY_PREFIX
+  #define CHOSEN_KEYPAD_DISPLAY_PREFIX                  "* 菜单"                                        // "* Menu"
+#endif
+#ifndef CHOSEN_KEYPAD_DISPLAY_SUFIX
+  #define CHOSEN_KEYPAD_DISPLAY_SUFIX                   "# 本键"                                   // "# This"
+#endif
+
+#ifndef CHOSEN_KEYPAD_0_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_0_DISPLAY_NAME                  "0 灯光"                                     // "0 Lights"
+#endif
+#ifndef CHOSEN_KEYPAD_1_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_1_DISPLAY_NAME                  "1 铃铛"                                      // "1 Bell"
+#endif
+#ifndef CHOSEN_KEYPAD_2_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_2_DISPLAY_NAME                  "2 喇叭"                                        // "2 Horn"
+#endif
+#ifndef CHOSEN_KEYPAD_3_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_3_DISPLAY_NAME                  "3 F3"                                          // "3 F3"
+#endif
+#ifndef CHOSEN_KEYPAD_4_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_4_DISPLAY_NAME                  "4 F4"                                          // "4 F4"
+#endif
+#ifndef CHOSEN_KEYPAD_5_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_5_DISPLAY_NAME                  "5 下一页"                                     // "5 Nxt Ttl"
+#endif
+#ifndef CHOSEN_KEYPAD_6_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_6_DISPLAY_NAME                  "6 速度倍率"                                    // "6 X Spd"
+#endif
+#ifndef CHOSEN_KEYPAD_7_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_7_DISPLAY_NAME                  "7 后退"                                      // "7 Rev"
+#endif
+#ifndef CHOSEN_KEYPAD_8_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_8_DISPLAY_NAME                  "8 紧急停车"                                     // "8 Estop"
+#endif
+#ifndef CHOSEN_KEYPAD_9_DISPLAY_NAME
+  #define CHOSEN_KEYPAD_9_DISPLAY_NAME                  "9 前进"                                    // "9 Fwd"
+#endif